Moved base64 encoding to utils directory, fixed some formatting
